+++ app/models/sheet.rb
@@ -249,7 +249,7 @@ class Sheet < ApplicationRecord
   end
 
   def original_attributes
-    ignore_attributes = %w(created_at updated_at authentication_token deleted successfully_validated)
+    ignore_attributes = %w(created_at updated_at authentication_token deleted successfully_validated initial_language_code)
     previous_changes.reject { |k, _v| ignore_attributes.include?(k.to_s) }.collect { |k, v| [k, v[0]] }
   end