+++ db/migrate/20180905151356_add_initial_language_code_to_sheets.rb
@@ -1,0 +1,5 @@
+class AddInitialLanguageCodeToSheets < ActiveRecord::Migration[5.2]
+  def change
+    add_column :sheets, :initial_language_code, :string
+  end
+end