+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 - **General Changes**
   - Update file inputs to match custom Bootstrap file inputs
+- **Export Changes**
+  - Initial language used for sheet data entry is now displayed in exports
 - **Search Changes**
   - Searches now includes better results, specifically for the variables
 - **Theme Changes**